+++ packages/web-tracing/src/types.ts
@@ -1,7 +1,7 @@
 import type { ContextManager, TextMapPropagator } from '@opentelemetry/api';
 import type { InstrumentationOption } from '@opentelemetry/instrumentation';
 import type { ResourceAttributes } from '@opentelemetry/resources';
-import type { SpanProcessor } from '@opentelemetry/sdk-trace-base';
+import type { SpanProcessor } from '@opentelemetry/sdk-trace-web';
 
 import type { API } from '@grafana/faro-web-sdk';