+++ packages/web-sdk/src/instrumentations/session/sessionManager/sessionManagerUtils.test.ts
@@ -97,6 +97,7 @@ describe('sessionManagerUtils', () => {
 
     const config = mockConfig({
       sessionTracking: {
+        enabled: true,
         onSessionChange: mockOnSessionChange,
       },
     });
@@ -160,9 +161,10 @@ describe('sessionManagerUtils', () => {
 
     const config = mockConfig({
       sessionTracking: {
+        enabled: true,
         onSessionChange: mockOnSessionChange,
+        session: currentSessionMeta,
       },
-      session: currentSessionMeta,
     });
 
     const faro = initializeFaro(config);
@@ -209,7 +211,7 @@ describe('sessionManagerUtils', () => {
 
   it('Takes session object and adds meta data.', () => {
     const config = mockConfig({});
-    initializeFaro(config);
+    const { api } = initializeFaro(config);
 
     const newSession: FaroUserSession = {
       lastActivity: 1,
@@ -259,10 +261,7 @@ describe('sessionManagerUtils', () => {
       },
     };
 
-    const config2 = mockConfig({
-      session: sessionMeta,
-    });
-    initializeFaro(config2);
+    api.setSession(sessionMeta);
 
     const sessionWithMetadata3 = addSessionMetadataToNextSession(newSession, previousSession);