+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next
 
-- Deps: remove unused `@opentelemetry/api-metrics` dependency (#401).
+- Deps (Web Tracing): remove unused `@opentelemetry/api-metrics` dependency (#401).
 
 ## 1.2.9